Potential fix for code scanning alert no. 50: Multiplication result converted to larger type
To fix the problem, we need to ensure that the multiplication is performed using the larger type (size_t) to avoid overflow. This can be achieved by casting one of the operands to size_t before performing the multiplication. Specifically, we will cast config.output.u.RGBA.stride to size_t before multiplying it by sub_image.height.

+++ libwebp-1.0.3/src/mux/anim_encode.c
@@ -1433,7 +1433,7 @@ static int DecodeFrameOntoCanvas(const WebPMuxFrameInfo* const frame,
   config.output.colorspace = MODE_BGRA;
   config.output.u.RGBA.rgba = (uint8_t*)sub_image.argb;
   config.output.u.RGBA.stride = sub_image.argb_stride * 4;
-  config.output.u.RGBA.size = config.output.u.RGBA.stride * sub_image.height;
+  config.output.u.RGBA.size = (size_t)config.output.u.RGBA.stride * sub_image.height;
 
   if (WebPDecode(image->bytes, image->size, &config) != VP8_STATUS_OK) {
     return 0;
